+++ gufsm/FSMVectorRunner.cc
@@ -2,7 +2,7 @@
  *  FSMVectorRunner.cc
  *  
  *  Created by René Hexel on 31/05/12.
- *  Copyright (c) 2011, 2012, 2013 Rene Hexel. All rights reserved.
+ *  Copyright (c) 2011, 2012, 2013, 2014 Rene Hexel. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -223,7 +223,7 @@ public:
                 if (index--)
                 {
                         if (!index) _content.clear();
-                        if (index >= _content.size())
+                        if (static_cast<size_t>(index) >= _content.size())
                                 _content.push_back(value);
                         else
                                 _content[index] = value;
@@ -256,7 +256,7 @@ public:
                 if (msg.getType() == WBMsg::TypeArray)
                 {
                         const vector<int> &vec = msg.getArrayValue();
-                        if (_index >= 0 && vec.size() > _index)
+                        if (_index >= 0 && vec.size() > static_cast<size_t>(_index))
                                 return vec[_index];
                         return 0;
                 }
